Small security update

+++ main.py
@@ -56,43 +56,52 @@ async def send_video_every():
 # Кнопка настройки
 @dp.message(F.text.lower() == "настройки")
 async def setting(message: types.Message) -> None:
-    await message.answer("Если вы хотите добавить или удалить канал из списка каналов, просто напишите его название в чат")
-    
+    if message.from_user.id == settings['user_wl']:
+        await message.answer("Если вы хотите добавить или удалить канал из списка каналов, просто напишите его название в чат")
+    else:
+        await message.answer(f"Привет, тебя нету в вайт листе, так что ты не можешь пользоваться ботом\nТвой id для добавления в white list: {message.from_user.id}")
+
+
 
 # Кнопка информации
 @dp.message(F.text.lower() == "информация")
 async def information(message: types.Message):
-    a = ""
-    for i in data:
-        a = a + f"\nhttps://www.youtube.com/channel/{i}"
-    await message.answer(f"Список каналов:{a}")
+    if message.from_user.id == settings['user_wl']:
+        a = ""
+        for i in data:
+            a = a + f"\nhttps://www.youtube.com/channel/{i}"
+        await message.answer(f"Список каналов:{a}")
+    else:
+        await message.answer(f"Привет, тебя нету в вайт листе, так что ты не можешь пользоваться ботом\nТвой id для добавления в white list: {message.from_user.id}")
 
 
 # Кнопка загрузки подкаста
 @dp.message(F.text.lower() == "загрузить подкаст")
 async def dowload_without(message: types.Message):
-    videos = scrapetube.get_channel(random.choice(data))
-    i = 0 
-    vid_list = []
-    for video in videos:
-        print(video['videoId'])
-        vid_list.append(video['videoId'])
-        i = i+1
-        if i == 10:
-            break
-    a = random.randint(1,99999999999)
-    yt = YouTube(f'https://www.youtube.com/watch?v={random.choice(vid_list)}')
-    stream = yt.streams.filter(only_audio=True).first()
-    stream.download(filename=f"{yt.title}.mp3")
-    audio = FSInputFile(f"{yt.title}.mp3")
-    await message.answer_audio(audio=audio, caption=f"Держи свой подкаст под названием {yt.title}!\n{yt.watch_url}")
-    os.remove(f"{yt.title}.mp3")
+    if message.from_user.id == settings['user_wl']:
+        videos = scrapetube.get_channel(random.choice(data))
+        i = 0 
+        vid_list = []
+        for video in videos:
+            print(video['videoId'])
+            vid_list.append(video['videoId'])
+            i = i+1
+            if i == 10:
+                break
+        a = random.randint(1,99999999999)
+        yt = YouTube(f'https://www.youtube.com/watch?v={random.choice(vid_list)}')
+        stream = yt.streams.filter(only_audio=True).first()
+        stream.download(filename=f"{yt.title}.mp3")
+        audio = FSInputFile(f"{yt.title}.mp3")
+        await message.answer_audio(audio=audio, caption=f"Держи свой подкаст под названием {yt.title}!\n{yt.watch_url}")
+        os.remove(f"{yt.title}.mp3")
+    else:
+        await message.answer(f"Привет, тебя нету в вайт листе, так что ты не можешь пользоваться ботом\nТвой id для добавления в white list: {message.from_user.id}")
 
 
 # /start
 @dp.message(Command("start"))
 async def cmd_start(message: types.Message):
-    print(message.from_user.id)
     if message.from_user.id == settings['user_wl']:
         kb = [
             [
@@ -134,22 +143,25 @@ async def main():
 # Функция при получении сообщения
 @dp.message()
 async def name(message: types.Message):
-    a = True
-    b = 0
-    channel_id = requests.get(f'https://www.googleapis.com/youtube/v3/search?part=id&q={message.text}&type=channel&key={settings["token_google"]}').json()['items'][0]['id']['channelId']
-    if len(channel_id)==24:
-        for i in data:
-            if channel_id == i:
-                data.pop(b)
-                a = False
-            b = b + 1
-        if a == True:
-            data.append(channel_id)
-            with open('data.json', 'w') as outfile:
-                json.dump(data, outfile, indent=4)
-            await message.answer(f"Добавил в список каналов на отправку подкастов:\nhttps://www.youtube.com/channel/{channel_id}\n\nЕсли вы хотите удалить его из списка, напишите название ещё раз")
-        else:
-            await message.answer(f"Данный канал удалён из списка каналов:\nhttps://www.youtube.com/channel/{channel_id}")
+    if message.from_user.id == settings['user_wl']:
+        a = True
+        b = 0
+        channel_id = requests.get(f'https://www.googleapis.com/youtube/v3/search?part=id&q={message.text}&type=channel&key={settings["token_google"]}').json()['items'][0]['id']['channelId']
+        if len(channel_id)==24:
+            for i in data:
+                if channel_id == i:
+                    data.pop(b)
+                    a = False
+                b = b + 1
+            if a == True:
+                data.append(channel_id)
+                with open('data.json', 'w') as outfile:
+                    json.dump(data, outfile, indent=4)
+                await message.answer(f"Добавил в список каналов на отправку подкастов:\nhttps://www.youtube.com/channel/{channel_id}\n\nЕсли вы хотите удалить его из списка, напишите название ещё раз")
+            else:
+                await message.answer(f"Данный канал удалён из списка каналов:\nhttps://www.youtube.com/channel/{channel_id}")
+    else:
+        await message.answer(f"Привет, тебя нету в вайт листе, так что ты не можешь пользоваться ботом\nТвой id для добавления в white list: {message.from_user.id}")
